Update SDK for Kinetis KW Family to Version 2.8.0

+++ devices/MKW36A4/drivers/fsl_tpm.c
@@ -2,7 +2,7 @@
  * Copyright (c) 2015, Freescale Semiconductor, Inc.
  * Copyright 2016-2017 NXP
  * All rights reserved.
- * 
+ *
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
@@ -147,7 +147,7 @@ status_t TPM_SetupPwm(TPM_Type *base,
     assert(numOfChnls);
     assert(srcClock_Hz);
 #if defined(FSL_FEATURE_TPM_HAS_COMBINE) && FSL_FEATURE_TPM_HAS_COMBINE
-    if(mode == kTPM_CombinedPwm)
+    if (mode == kTPM_CombinedPwm)
     {
         assert(FSL_FEATURE_TPM_COMBINE_HAS_EFFECTn(base));
     }
@@ -160,7 +160,7 @@ status_t TPM_SetupPwm(TPM_Type *base,
 
 #if defined(FSL_FEATURE_TPM_HAS_QDCTRL) && FSL_FEATURE_TPM_HAS_QDCTRL
     /* The TPM's QDCTRL register required to be effective */
-    if( FSL_FEATURE_TPM_QDCTRL_HAS_EFFECTn(base) )
+    if (FSL_FEATURE_TPM_QDCTRL_HAS_EFFECTn(base))
     {
         /* Clear quadrature Decoder mode because in quadrature Decoder mode PWM doesn't operate*/
         base->QDCTRL &= ~TPM_QDCTRL_QUADEN_MASK;
@@ -346,7 +346,7 @@ void TPM_UpdatePwmDutycycle(TPM_Type *base,
 {
     assert(chnlNumber < FSL_FEATURE_TPM_CHANNEL_COUNTn(base));
 #if defined(FSL_FEATURE_TPM_HAS_COMBINE) && FSL_FEATURE_TPM_HAS_COMBINE
-    if(currentPwmMode == kTPM_CombinedPwm)
+    if (currentPwmMode == kTPM_CombinedPwm)
     {
         assert(FSL_FEATURE_TPM_COMBINE_HAS_EFFECTn(base));
     }
@@ -391,7 +391,7 @@ void TPM_UpdatePwmDutycycle(TPM_Type *base,
 
 void TPM_UpdateChnlEdgeLevelSelect(TPM_Type *base, tpm_chnl_t chnlNumber, uint8_t level)
 {
-    assert(chnlNumber < FSL_FEATURE_TPM_CHANNEL_COUNTn(base));
+    assert((uint8_t)chnlNumber < FSL_FEATURE_TPM_CHANNEL_COUNTn(base));
 
     uint32_t reg = base->CONTROLS[chnlNumber].CnSC & ~(TPM_CnSC_CHF_MASK);
 
@@ -421,11 +421,11 @@ void TPM_UpdateChnlEdgeLevelSelect(TPM_Type *base, tpm_chnl_t chnlNumber, uint8_
 
 void TPM_SetupInputCapture(TPM_Type *base, tpm_chnl_t chnlNumber, tpm_input_capture_edge_t captureMode)
 {
-    assert(chnlNumber < FSL_FEATURE_TPM_CHANNEL_COUNTn(base));
+    assert((uint8_t)chnlNumber < FSL_FEATURE_TPM_CHANNEL_COUNTn(base));
 
 #if defined(FSL_FEATURE_TPM_HAS_QDCTRL) && FSL_FEATURE_TPM_HAS_QDCTRL
     /* The TPM's QDCTRL register required to be effective */
-    if( FSL_FEATURE_TPM_QDCTRL_HAS_EFFECTn(base) )
+    if (FSL_FEATURE_TPM_QDCTRL_HAS_EFFECTn(base))
     {
         /* Clear quadrature Decoder mode for channel 0 or 1*/
         if ((chnlNumber == 0) || (chnlNumber == 1))
@@ -436,8 +436,8 @@ void TPM_SetupInputCapture(TPM_Type *base, tpm_chnl_t chnlNumber, tpm_input_capt
 #endif
 
 #if defined(FSL_FEATURE_TPM_HAS_COMBINE) && FSL_FEATURE_TPM_HAS_COMBINE
-        /* The TPM's COMBINE register required to be effective */
-    if( FSL_FEATURE_TPM_COMBINE_HAS_EFFECTn(base) )
+    /* The TPM's COMBINE register required to be effective */
+    if (FSL_FEATURE_TPM_COMBINE_HAS_EFFECTn(base))
     {
         /* Clear the combine bit for chnlNumber */
         base->COMBINE &= ~(1U << TPM_COMBINE_SHIFT * (chnlNumber / 2));
@@ -469,11 +469,11 @@ void TPM_SetupOutputCompare(TPM_Type *base,
                             tpm_output_compare_mode_t compareMode,
                             uint32_t compareValue)
 {
-    assert(chnlNumber < FSL_FEATURE_TPM_CHANNEL_COUNTn(base));
+    assert((uint8_t)chnlNumber < FSL_FEATURE_TPM_CHANNEL_COUNTn(base));
 
 #if defined(FSL_FEATURE_TPM_HAS_QDCTRL) && FSL_FEATURE_TPM_HAS_QDCTRL
     /* The TPM's QDCTRL register required to be effective */
-    if( FSL_FEATURE_TPM_QDCTRL_HAS_EFFECTn(base) )
+    if (FSL_FEATURE_TPM_QDCTRL_HAS_EFFECTn(base))
     {
         /* Clear quadrature Decoder mode for channel 0 or 1 */
         if ((chnlNumber == 0) || (chnlNumber == 1))
@@ -520,7 +520,7 @@ void TPM_SetupDualEdgeCapture(TPM_Type *base,
 
 #if defined(FSL_FEATURE_TPM_HAS_QDCTRL) && FSL_FEATURE_TPM_HAS_QDCTRL
     /* The TPM's QDCTRL register required to be effective */
-    if( FSL_FEATURE_TPM_QDCTRL_HAS_EFFECTn(base) )
+    if (FSL_FEATURE_TPM_QDCTRL_HAS_EFFECTn(base))
     {
         /* Clear quadrature Decoder mode for channel 0 or 1*/
         if (chnlPairNumber == 0)
